fix:미션 난이도 수정 및 로직 수정

+++ src/main/java/com/dku/mentoring/register/service/RegisterService.java
@@ -9,6 +9,7 @@ import com.dku.mentoring.mission.model.entity.MissionBonus;
 import com.dku.mentoring.mission.repository.MissionBonusRepository;
 import com.dku.mentoring.mission.repository.MissionRepository;
 import com.dku.mentoring.register.exception.NoRightToAccessException;
+import com.dku.mentoring.register.exception.RegisterAlreadyCompleteException;
 import com.dku.mentoring.register.exception.RegisterNotfoundException;
 import com.dku.mentoring.register.model.dto.list.SummarizedRegisterDto;
 import com.dku.mentoring.register.model.dto.request.AdminApproveRequestDto;
@@ -16,13 +17,16 @@ import com.dku.mentoring.register.model.dto.request.RegisterRequestDto;
 import com.dku.mentoring.register.model.dto.response.SingleRegisterResponseDto;
 import com.dku.mentoring.register.model.entity.Register;
 import com.dku.mentoring.register.model.entity.RegisterFile;
+import com.dku.mentoring.register.model.entity.RegisterStatus;
 import com.dku.mentoring.register.repository.RegisterRepository;
 import com.dku.mentoring.user.entity.User;
 import com.dku.mentoring.user.exception.UserNotFoundException;
 import com.dku.mentoring.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -83,17 +87,20 @@ public class RegisterService {
      * @return 페이징된 목록
      */
     public Page<SummarizedRegisterDto> getRegisters(Pageable pageable) {
+        pageable = makePageableDesc(pageable);
         Page<Register> registers = registerRepository.findAllRegisters(pageable);
         return registers.map(SummarizedRegisterDto::new);
     }
 
     /**
-     * 사용자가 등록 글 전체 조회
+     * 사용자가 등록한 글 전체 조회
      *
      */
     public Page<SummarizedRegisterDto> getRegistersByUser(HttpServletRequest request, Pageable pageable) {
         User user = getMemberFromRequest(request);
         Long userId = user.getId();
+
+        pageable = makePageableDesc(pageable);
 
         Page<Register> registers = registerRepository.findByUserId(userId, pageable);
         return registers.map(SummarizedRegisterDto::new);
@@ -127,6 +134,9 @@ public class RegisterService {
         if(!register.getUser().getId().equals(user.getId())) {
             throw new NoRightToAccessException("해당 권한이 없습니다.");
         }
+        if(register.getStatus().equals(RegisterStatus.COMPLETE)){
+            throw new RegisterAlreadyCompleteException("이미 완료된 글입니다.");
+        }
         registerRepository.delete(register);
     }
 
@@ -138,6 +148,9 @@ public class RegisterService {
     @Transactional
     public void approveRegister(Long registerId, HttpServletRequest request, AdminApproveRequestDto dto) {
         Register register = registerRepository.findById(registerId).orElseThrow(() -> new RegisterNotfoundException("해당 글이 없습니다."));
+        if(register.getStatus().equals(RegisterStatus.COMPLETE)){
+            throw new RegisterAlreadyCompleteException("이미 완료된 글입니다.");
+        }
         User user = getMemberFromRequest(request);
         if(user.getRoles().stream().noneMatch(role -> role.getRolename().equals("ROLE_ADMIN"))) {
             throw new NoRightToAccessException("해당 권한이 없습니다.");
@@ -156,5 +169,14 @@ public class RegisterService {
         User user = userRepository.findByStudentId(studentId).orElseThrow(UserNotFoundException::new);
 
         return user;
+    }
+
+    /**
+     * 페이징 방법을 내림차순으로 변경
+     */
+    private static Pageable makePageableDesc(Pageable pageable) {
+        Sort sortByCreatedAtDesc = Sort.by("createdAt").descending();
+        pageable = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(), sortByCreatedAtDesc);
+        return pageable;
     }
 }
